+++ tools/testing/selftests/bpf/progs/netif_receive_skb.c
@@ -9,7 +9,7 @@
 
 #include <errno.h>
 
-long ret = 0;
+long long ret = 0;
 int num_subtests = 0;
 int ran_subtests = 0;
 bool skip = false;
