+++ tools/testing/selftests/bpf/progs/test_perf_skip.c
@@ -3,7 +3,7 @@
 #include <bpf/bpf_helpers.h>
 #include <bpf/bpf_tracing.h>
 
-uintptr_t ip;
+u32 ip;
 
 SEC("perf_event")
 int handler(struct bpf_perf_event_data *data)
