+++ tools/testing/selftests/bpf/progs/test_ringbuf_n.c
@@ -22,7 +22,7 @@ struct {
 } ringbuf SEC(".maps");
 
 int pid = 0;
-long value = 0;
+long long value = 0;
 
 SEC("fentry/" SYS_PREFIX "sys_getpgid")
 int test_ringbuf_n(void *ctx)
