+++ tools/testing/selftests/bpf/progs/verifier_global_subprogs.c
@@ -106,7 +106,7 @@ int unguarded_unsupp_global_called(void)
 	return global_unsupp(&x);
 }
 
-long stack[128];
+long long stack[128];
 
 __weak int subprog_nullable_ptr_bad(int *p)
 {
